+++ Generals/Code/GameEngine/Source/Common/RTS/SpecialPower.cpp
@@ -282,7 +282,7 @@ const SpecialPowerTemplate *SpecialPowerStore::getSpecialPowerTemplateByIndex( U
 Int SpecialPowerStore::getNumSpecialPowers( void )
 {
 
-	return m_specialPowerTemplates.size();
+	return (int)m_specialPowerTemplates.size();
 
 }  // end getNumSpecialPowers
